+++ cmd/livesim2/app/handler_index.go
@@ -41,4 +41,5 @@ func (s *Server) favIconFunc(w http.ResponseWriter, r *http.Request) {
 // optionsHandlerFunc provides the allowed methods.
 func (s *Server) optionsHandlerFunc(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Allow", "OPTIONS, GET, HEAD")
+	w.WriteHeader(http.StatusNoContent)
 }